refactor(studio): split KV browser + fix bulk-delete failure handling
Addresses thermo review findings on the KV browser:

- Split the 1000-line kv-browser.tsx into sibling files per the
  data/ and storage/ pattern: kv-fields.ts (pure helpers +
  buildKvPutOptions), kv-form-fields.tsx (shared TTL/metadata
  fields), kv-value-editor.tsx, kv-create-form.tsx, kv-key-list.tsx,
  and a slim kv-browser.tsx shell.
- De-duplicate the create form and value editor: they now share the
  field components and a single buildKvPutOptions payload contract.
- Fix bulk delete swallowing failures: use Promise.allSettled, prune
  only keys that actually deleted, keep failed keys selected, and
  surface a "Failed to delete N keys" error (was: removed all rows +
  no error on any failure).
- Validate TTL >= 60s client-side (Cloudflare KV's minimum) instead
  of round-tripping to a server error.
- Drop the dead clearBulk reducer case.
- Add tests for the partial-failure and TTL-minimum paths (10 total).
