+++ packages/studio/src/locales/en.ts
@@ -838,8 +838,10 @@ const MESSAGE_IDS = [
     "Copy key",
     "Copy value",
     "Create key",
+    "Failed to delete {count} keys.",
     "Format JSON",
     "Key name",
+    "TTL must be a whole number of at least 60 seconds.",
     "Leave blank to keep the current expiration.",
     "Metadata (JSON)",
     "Metadata must be valid JSON.",
